+++ members/nullnet-client/src/commands/mod.rs
@@ -113,6 +113,13 @@ pub(crate) struct RtNetLinkHandle {
 }
 
 impl RtNetLinkHandle {
+    #[cfg(test)]
+    pub(crate) fn disconnected() -> Self {
+        let (connection, handle, _) = new_connection().unwrap();
+        drop(connection);
+        Self { handle }
+    }
+
     pub(crate) fn new() -> Result<Self, Error> {
         let (rtnetlink_conn, rtnetlink_handle, _) = new_connection().handle_err(location!())?;
         tokio::spawn(rtnetlink_conn);